update snapshots and specs
